+++ src/shared/content-shared/content-lab/content-lab-button/content-lab-button.component.ts
@@ -1,33 +1,15 @@
-import {Component, EventEmitter, Input, OnDestroy, Output, ViewChild} from '@angular/core';
-import {
-    KalturaClient,
-    KalturaEntryStatus,
-    KalturaMediaEntry,
-    KalturaMediaType,
-    KalturaPlaylist,
-    KalturaQuizOutputType,
-    QuizGetUrlAction
-} from 'kaltura-ngx-client';
+import {Component, Input, OnDestroy} from '@angular/core';
+import {KalturaEntryStatus, KalturaMediaEntry} from 'kaltura-ngx-client';
 import {cancelOnDestroy, tag} from '@kaltura-ng/kaltura-common';
-import {AppAuthentication, AppBootstrap, ApplicationType, BrowserService} from 'app-shared/kmc-shell';
-import {PreviewAndEmbedEvent} from 'app-shared/kmc-shared/events';
-import {ISubscription} from 'rxjs/Subscription';
-import {PubSubServiceType, UnisphereElementBaseType} from '@unisphere/runtime';
-import {Router} from '@angular/router';
+import {AppBootstrap} from 'app-shared/kmc-shell';
 import { ChangeDetectorRef } from '@angular/core';
-import {AppEventsService} from 'app-shared/kmc-shared';
-
-import {PopupWidgetComponent} from '@kaltura-ng/kaltura-ui';
 import {KalturaLogger} from '@kaltura-ng/kaltura-logger';
-import {AppLocalization} from '@kaltura-ng/mc-shared';
-import {Observable} from 'rxjs';
-import {throwError as ObservableThrowError} from 'rxjs/internal/observable/throwError';
 
 @Component({
     selector: 'k-content-lab-btn',
     templateUrl: './content-lab-button.component.html',
     styleUrls: ['./content-lab-button.component.scss'],
-    providers: [KalturaLogger.createLogger('ContentLabButtonComponent')]
+    providers: [KalturaLogger.createLogger('ContentLabBtnComponent')]
 })
 export class ContentLabBtnComponent implements OnDestroy {
     @Input() set entryId(value: string) {
@@ -40,14 +22,7 @@ export class ContentLabBtnComponent implements OnDestroy {
     @Input() entryType: number;
     @Input() eventSessionContextId: string;
     @Input() responsive: boolean;
-    @Input() entryStore: any;
 
-    @Output() onAction = new EventEmitter<string>();
-
-    @ViewChild('clipAndTrim', { static: true }) _clipAndTrim: PopupWidgetComponent;
-    @ViewChild('bulkActionsPopup', { static: true }) _bulkActionsPopup: PopupWidgetComponent;
-
-    private unisphereCallbackUnsubscribe: ISubscription;
     private unisphereRuntime: any = null;
     private _unsubscribePartnerCheck: () => void;
     private _destroyed = false;
@@ -58,27 +33,15 @@ export class ContentLabBtnComponent implements OnDestroy {
     public reason = '';
     public _entryId: string;
 
-
-    constructor(private _bootstrapService: AppBootstrap,
-                private _router: Router,
-                private cdr: ChangeDetectorRef,
-                private _kalturaServerClient: KalturaClient,
-                private _appEvents: AppEventsService,
-                private _browserService: BrowserService,
-                private _appLocalization: AppLocalization,
-                private _logger: KalturaLogger,
-                private _appAuthentication: AppAuthentication) {
+    constructor(private _bootstrapService: AppBootstrap, private cdr: ChangeDetectorRef) {
     }
 
     private initializeUnisphereRuntime() {
         this._bootstrapService.unisphereWorkspace$
             .pipe(cancelOnDestroy(this))
             .subscribe(unisphereWorkspace => {
-
                     if (unisphereWorkspace) {
-
                         this.unisphereRuntime = unisphereWorkspace.getRuntime('unisphere.widget.content-lab', 'application');
-
                         if (this.unisphereRuntime) {
                             this._unsubscribePartnerCheck = this.unisphereRuntime.partnerChecks.onChanges((data) => {
                                 if (data.status === 'loaded') {
@@ -102,6 +65,7 @@ export class ContentLabBtnComponent implements OnDestroy {
                                                 this.loading = false;
                                                 this.disabled = true;
                                                 this.reason = 'error';
+                                                this.cdr.detectChanges();
                                             }
                                         )
                                     } else {
@@ -124,71 +88,6 @@ export class ContentLabBtnComponent implements OnDestroy {
                             }, {replayLastValue: true})
 
                         }
-
-                        unisphereWorkspace.getService<PubSubServiceType>('unisphere.service.pub-sub')?.subscribe('unisphere.event.module.content-lab.message-host-app', (data) => {
-                            const { action, entry } = data.payload;
-                            if (entry.id === this._entryId) {
-                                switch (action) {
-                                    case 'entry':
-                                        // navigate to entry
-                                        this.unisphereRuntime?.closeWidget(); // close widget
-                                        document.body.style.overflowY = "auto";
-                                        if (this.entryStore) {
-                                            this.entryStore.openEntry(new KalturaMediaEntry(entry));
-                                        } else {
-                                            this._router.navigateByUrl(`/content/entries/entry/${entry.id}/metadata`);
-                                        }
-                                        break;
-                                    case 'playlist':
-                                        // navigate to playlist metadata tab
-                                        this.unisphereRuntime?.closeWidget(); // close widget
-                                        document.body.style.overflowY = "auto";
-                                        this._router.navigateByUrl(`/content/playlists/playlist/${entry.id}/metadata`);
-                                        break;
-                                    case 'editPlaylist':
-                                        // navigate to playlist content tb
-                                        this.unisphereRuntime?.closeWidget(); // close widget
-                                        document.body.style.overflowY = "auto";
-                                        this._router.navigateByUrl(`/content/playlists/playlist/${entry.id}/content`);
-                                        break;
-                                    case 'download':
-                                        // download entry
-                                        const downloadUrl = entry.downloadUrl.indexOf('/ks/') === -1 ? `${entry.downloadUrl}/ks/${this._appAuthentication.appUser.ks}` : entry.downloadUrl;
-                                        this._browserService.openLink(downloadUrl);
-                                        break;
-                                    case 'share':
-                                        // open share & embed for entry
-                                        this.unisphereRuntime?.closeWidget(); // close widget
-                                        document.body.style.overflowY = "auto";
-                                        this._appEvents.publish(new PreviewAndEmbedEvent(new KalturaMediaEntry(entry)));
-                                        break;
-                                    case 'sharePlaylist':
-                                        // open share & embed for playlist
-                                        this.unisphereRuntime?.closeWidget(); // close widget
-                                        document.body.style.overflowY = "auto";
-                                        this._appEvents.publish(new PreviewAndEmbedEvent(new KalturaPlaylist(entry)));
-                                        break;
-                                    case 'editQuiz':
-                                        // edit entry
-                                        this._contentLabSelectedQuiz = new KalturaMediaEntry(entry);
-                                        this.unisphereRuntime?.closeWidget();
-                                        document.body.style.overflowY = "auto";
-                                        // this._clipAndTrim.open();
-                                        this.onAction.emit('editQuiz');
-                                        break;
-                                    case 'downloadQuiz':
-                                        // download questions list
-                                        this._downloadPretest(entry.id)
-                                        break;
-                                    case 'updateMetadata':
-                                        // update metadata
-                                        // this._entryStore.reloadEntry(); TODO: send action
-                                        break;
-                                    default:
-                                        break;
-                                }
-                            }
-                        })
                     }
                 },
                 error => {
@@ -210,34 +109,4 @@ export class ContentLabBtnComponent implements OnDestroy {
         }
     }
 
-    private downloadPretest(entryId: string): Observable<string> {
-        if (!entryId) {
-            return ObservableThrowError('missing entryId argument');
-        }
-        return this._kalturaServerClient
-            .request(new QuizGetUrlAction({ entryId, quizOutputType: KalturaQuizOutputType.pdf }))
-    }
-
-    private _downloadPretest(entryId: string): void {
-        if (!entryId) {
-            this._logger.info('EntryId is not defined. Abort action');
-            return;
-        }
-        this.downloadPretest(entryId)
-            .pipe(
-                tag('block-shell'),
-                cancelOnDestroy(this)
-            )
-            .subscribe(
-                (url) => {
-                    this._browserService.openLink(url);
-                },
-                error => {
-                    this._browserService.alert({
-                        header: this._appLocalization.get('app.common.error'),
-                        message: error.message
-                    });
-                }
-            );
-    }
 }
